requirements.txt
jinja2
jsonschema
PyYAML